fix: change modal
