+++ src/components/Modal/components/PhoneReplaceContent/index.ts
@@ -1,0 +1,1 @@
+export { PhoneReplaceContent } from './PhoneReplaceContent';
